Use Michiel's init-script and plain maria-db image
This way, usage of the prototype-db does not require a docker login.
Nor do we need to maintain our own dockerfile for the ampersand DB.

Note, the extension of the added file in db-init-scripts is rather
important. In an earlier attempt we had a filename with a .d extension,
which gets ignored. Similarly, when renamed as .sh it will be executed
with bash.

Interestingly, the script only gets run when the volume used to persist
the database does not exist. Moreover, that volume stores the
permissions of the DB. Hence if you have wrong code here, but a volume
that is right, you will not notice the wrong code.

The current code has been confirmed on two separate PCs, both by
clearing the volumes.